random stuff
In this commit, I started working on the project. I wrote a decent README file, I wrote a basic HTML page and some CSS to show the general visual style of the app, added a gitignore, and introduced this very poggers concept of an announcements.js file which people hosting this app can use for announcements. In the future, I intend to include holiday themes in this file for fun.