finished every class but DailyLog
(Day is called DailyLog now. might change it back later.) i'm feeling good today. i managed to finally convince myself that i'm not a terrible person and now it's like my brain has 5x as much free ram and cpu time since i'm not occupied with feeling guilty and ashamed all the time. anyway, the code has comments as usual. there is a system now where fluids can be set to be permanently displayed or not. the ones that are not set to be permanent are not saved until they are actually used. fluids are saved so that if a fluid is ever removed from custom.js, it will still show up properly when viewing previous days' logs. now how might i deal with users changing time zones...?